Move calculators into calculators directory
