+++ src/calculators/utils/compute-compound-interest.js
