+++ src/calculators/utils/compute-compound-interest.test.js
