+++ src/calculators/utils/compute-cycle.js
