+++ src/calculators/utils/normalize-portfolio.js
