+++ src/calculators/utils/normalize-portfolio.test.js
