+++ src/calculators/utils/spending.js
